+++ Sources/KeyboardKit/Previews/SecondaryCalloutActionProvider+Preview.swift
@@ -1,0 +1,22 @@
+//
+//  StandardSecondaryCalloutActionProvider+Preview.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-02-01.
+//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This class can be used when previewing keyboard views.
+ */
+public class PreviewSecondaryCalloutActionProvider {
+    
+    private init() {}
+    
+    static var preview: SecondaryCalloutActionProvider {
+        StandardSecondaryCalloutActionProvider(
+            context: FakeKeyboardContext())
+    }
+}